Create simulation package
